First Changes
Change the jump to the isInvalid branch to a new Correct Message branch to output the right messages. Store the length of the string into an a register so to be used in the subprogram. I then deleted my doLoop and in its place began the code for the CM branch.