Continue initial changes
I finished coding the CorrectMessage branch by comparing the length to all valid lengths and printing out the IsInvalid message as this branch is only entered if a space is detected. Since I deleted my DoLoop, I replaced that with the call to conversion function which will set up my sub program.