Allocating memory
I then allocated new memory on the stack for return address and the contents in $s3. I then coded my base case which states if the length is 0 then end the recursion. I made sure to start at the end of the string as my first power equals to 1, so I will use my exponentiated base working "backwards"